cache source code and go build cache (#134)
* remove softlink to gopath and add nfs git cache

* fix checkout dir

* Add race test (#135)

* revert add race test , race test will run with a new separate job (#136)

* Add race test

* Revert "Add race test"

This reverts commit 234a3db868f6152d047ffa251c451802f40249ab.

* Revert "Add race test (#135)" (#137)

This reverts commit f5ec4b8cd9794ad8f43fd3bb76804073895348e0.

* update tidb related script to use cache file

* copy code from nfs cacaed dir (#138)

* add pod selector for tics test

* replace pod test_go_heavy to test_go (#140)

* copy code from nfs cacaed dir

* pod replace: test_go_heavy to test_go

* replace test_go_heavy to test_go (#141)

* fix tidb_ghpr_build tidb plugin dir

* fix tidb coverage test script

* Copy tidb code from nfs cache  (#142)

* copy code from nfs cacaed dir

* pod replace: test_go_heavy to test_go

* replace pod test_go_heavy to test_go

* replace test_go_heavy to test_go

* copy tidb code from nfs cache

* Remove pod idleMinutes to destroy tidb-test-tics pod when job finished (#144)

* copy code from nfs cacaed dir

* pod replace: test_go_heavy to test_go

* replace pod test_go_heavy to test_go

* replace test_go_heavy to test_go

* copy tidb code from nfs cache

* remove pod idleMinutes to destroy tidb-test-tics pod when job finished

* Fix clone code too long, Add get code from fileserver if nfs server down (#147)

* copy code from nfs cacaed dir

* pod replace: test_go_heavy to test_go

* replace pod test_go_heavy to test_go

* replace test_go_heavy to test_go

* copy tidb code from nfs cache

* remove pod idleMinutes to destroy tidb-test-tics pod when job finished

* Fix clone too long, Add get code from fileserver if nfs server down

* Add race test (#149)

* fix clone code takes long time for the rest tidb scripts (#150)

* fix clone code takes long time for the rest tidb scripts

* fix sh script dirs


* Fix tar dir not exist error (#151)

* Fix tar dir not exist error

* test nfs src.tar exist

* fix fileserver tar no such dir err for the rest of tidb scripts (#152)

* Parallel builds to reduce time (#153)

* parallel build tasks

* parallel build tasks

* tidb linux amd all build job run in parallel (#154)

* Fix bug when build tidb & plugins, tidb src not exist (#156)

* surround the deleteDir with node (#157)

* Make amd build in parallel, tidb get code from nfs-cache, fix tiflash clone error (#158)

* make arm64 build in parallel (#161)

* Cache code and go module (#168)

* add pipeline: update code & cache go module

* debug update code: use relative path to update softlink

* remove some dir step incase of too many @tmp dir

* remove debug info

* debug update code by step checkout scm

* fix bug use varible in url

* specify branch master

* remove container golang

* fix indent

* fix branch error

* update private repo

* hotfix release pipeline (#165)

* fix clone code takes long time for the rest tidb scripts

* fix sh script dirs

* add script to release hotfix

* add macos script to release hotfix

* fix some components binary name do not match

* add force_rebuild trigger to adjust hotfix release

* merge same output path rule

* add force_rebuild


* add tics ci pipeline (#171)

* add br ghpr pipeline (#172)

* fix sync-ci analysis-res is null (#176)


* make all build in mac to parallel (#177)

* add centos7_golang-1.16 Dockerfile (#178)


* jenkins: split unit and leak test to accelerate the speed (#173)

* jenkins: split unit and leak test to accelerate the speed

* fix bug: logic error, skip build if build cache existed (#181)

* tikv use test-slave-with-memvolume (#183)

* Patch tikv fix slave (#185)

* tikv use test-slave-with-memvolume

* Disable incremental build

* Patch2 darwin build parallel (#180)

* make all build in mac to parallel

* fix bug: correct function name

* fix bug: correct logic error, if buildcache exist then skip build this time

* fix bug: add node with label mac

* debug: make tiflash build in mac-i5

* fix conflict,ready to merge master (#186)

* Update release-check.groovy (#148)

* fix clone code takes long time for the rest tidb scripts

* fix sh script dirs

* make tidb_title_check more strict (#155)

* Update build-ucloud-image.groovy (#163)

* Update tikv_ghpr_build_release.groovy (#164)

* add script to release hotfix

* add macos script to release hotfix

* fix some components binary name do not match

* fix aborted job update comment (#167)


* add force_rebuild trigger to adjust hotfix release

* merge same output path rule

* add force_rebuild

* add pipeline to release tiup hotfix (#169)

* add pipeline to release tiup hotfix

* archive multi branch pipeline to github (#179)


* Disable  incremental build (#184)

set `export CARGO_INCREMENTAL=0` to disable incremental build

+++ jenkins/pipelines/ci/tidb/tidb_ghpr_check.groovy
@@ -43,36 +43,8 @@ try {
                 sh "whoami && go version"
             }
 
-            // update cache
-            dir("/home/jenkins/agent/git/tidb") {
-                if (sh(returnStatus: true, script: '[ -d .git ] && [ -f Makefile ] && git rev-parse --git-dir > /dev/null 2>&1') != 0) {
-                    deleteDir()
-                }
-                if (!fileExists("/home/jenkins/agent/git/tidb/Makefile")) {
-                    dir("/home/jenkins/agent/git") {
-                        sh """
-                            rm -rf tidb.tar.gz
-                            rm -rf tidb
-                            wget ${FILE_SERVER_URL}/download/source/tidb.tar.gz
-                            tar xvf tidb.tar.gz
-                        """
-                    }
-                }
-                dir("/home/jenkins/agent/git/tidb") {
-                    try {
-                        checkout changelog: false, poll: false, scm: [$class: 'GitSCM', branches: [[name: 'master']], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'PruneStaleBranch'], [$class: 'CleanBeforeCheckout'], [$class: 'CloneOption', timeout: 120]], submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'github-sre-bot-ssh', refspec: specStr, url: 'git@github.com:pingcap/tidb.git']]]
-                    } catch (error) {
-                        retry(2) {
-                            echo "checkout failed, retry.."
-                            sleep 2
-                            if (sh(returnStatus: true, script: '[ -d .git ] && [ -f Makefile ] && git rev-parse --git-dir > /dev/null 2>&1') != 0) {
-                                deleteDir()
-                            }
-                            checkout changelog: false, poll: false, scm: [$class: 'GitSCM', branches: [[name: 'master']], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'PruneStaleBranch'], [$class: 'CleanBeforeCheckout']], submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'github-sre-bot-ssh', refspec: specStr, url: 'git@github.com:pingcap/tidb.git']]]
-                        }
-                    }
-                }
-
+            // ensure golangci-lint tool exist
+            dir("/home/jenkins/agent/git/tools") {
                 if (!fileExists("/home/jenkins/agent/git/tools/bin/golangci-lint")) {
                     container("golang") {
                         dir("/home/jenkins/agent/git/tools/") {
@@ -83,32 +55,77 @@ try {
                     }
                 }
             }
+            // update code
+            dir("/home/jenkins/agent/code-archive") {
+                // delete to clean workspace in case of agent pod reused lead to conflict.
+                deleteDir()
+                // copy code from nfs cache
+                container("golang") {
+                    if(fileExists("/nfs/cache/git-test/src-tidb.tar.gz")){
+                        timeout(5) {
+                            sh """
+                                cp -R /nfs/cache/git-test/src-tidb.tar.gz*  ./
+                                mkdir -p ${ws}/go/src/github.com/pingcap/tidb
+                                tar -xzf src-tidb.tar.gz -C ${ws}/go/src/github.com/pingcap/tidb --strip-components=1
+                            """
+                        }
+                    }
+                }
+                dir("${ws}/go/src/github.com/pingcap/tidb") {
+                    if (sh(returnStatus: true, script: '[ -d .git ] && [ -f Makefile ] && git rev-parse --git-dir > /dev/null 2>&1') != 0) {
+                        echo "Not a valid git folder: ${ws}/go/src/github.com/pingcap/tidb"
+                        echo "Clean dir then get tidb src code from fileserver"
+                        deleteDir()
+                    }
+                    if(!fileExists("${ws}/go/src/github.com/pingcap/tidb/Makefile")) {
+                        dir("${ws}/go/src/github.com/pingcap/tidb") {
+                            sh """
+                                rm -rf /home/jenkins/agent/code-archive/tidb.tar.gz
+                                rm -rf /home/jenkins/agent/code-archive/tidb
+                                wget -O /home/jenkins/agent/code-archive/tidb.tar.gz  ${FILE_SERVER_URL}/download/source/tidb.tar.gz -q --show-progress
+                                tar -xzf /home/jenkins/agent/code-archive/tidb.tar.gz -C ./ --strip-components=1
+                            """
+                        }
+                    }
+                    try {
+                        checkout changelog: false, poll: false, scm: [$class: 'GitSCM', branches: [[name: 'master']], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'PruneStaleBranch'], [$class: 'CleanBeforeCheckout'], [$class: 'CloneOption', timeout: 2]], submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'github-sre-bot-ssh', refspec: specStr, url: 'git@github.com:pingcap/tidb.git']]]
+                    }   catch (info) {
+                            retry(2) {
+                                echo "checkout failed, retry.."
+                                sleep 5
+                                if (sh(returnStatus: true, script: '[ -d .git ] && [ -f Makefile ] && git rev-parse --git-dir > /dev/null 2>&1') != 0) {
+                                    deleteDir()
+                                }
+                                // if checkout one pr failed, we fallback to fetch all thre pr data
+                                checkout changelog: false, poll: false, scm: [$class: 'GitSCM', branches: [[name: 'master']], doGenerateSubmoduleConfigurations: false, extensions: [[$class: 'PruneStaleBranch'], [$class: 'CleanBeforeCheckout']], submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'github-sre-bot-ssh', refspec: specStr, url: 'git@github.com:pingcap/tidb.git']]]
+                            }
+                        }
+                    sh "git checkout -f ${ghprbActualCommit}"
+                }
+            }
         }
 
         stage("Build & Test") {
             container("golang") {
                 dir("go/src/github.com/pingcap/tidb") {
-                    deleteDir()
                     timeout(30) {
                         sh """
-                        cp -R /home/jenkins/agent/git/tidb/. ./
                         mkdir -p tools/bin
                         cp /home/jenkins/agent/git/tools/bin/golangci-lint tools/bin/
                         git checkout -f ${ghprbActualCommit}
                         ls -al tools/bin || true
                         # GOPROXY=http://goproxy.pingcap.net
-                        mkdir -p \$GOPATH/pkg/mod && mkdir -p ${ws}/go/pkg && ln -sfT \$GOPATH/pkg/mod ${ws}/go/pkg/mod
                         """
                     }
                     try {
                         if (ghprbTargetBranch == "master") {
                             def builds = [:]
                             builds["check"] = {
-                                sh "GOPATH=${ws}/go make check"
+                                sh "make check"
                             }
                             builds["test_part_1"] = {
                                 try {
-                                    sh "GOPATH=${ws}/go make test_part_1"
+                                    sh "make test_part_1"
                                 } catch (err) {
                                     throw err
                                 } finally {
@@ -117,7 +134,7 @@ try {
                             }
                             parallel builds
                         } else {
-                            sh "GOPATH=${ws}/go make dev"
+                            sh "make dev"
                         }
                     } catch (err) {
                         throw err
@@ -180,7 +197,7 @@ catch (Exception e) {
 stage("upload status") {
     node {
         println currentBuild.result
-        sh """curl --connect-timeout 2 --max-time 4 -d '{"job":"$JOB_NAME","id":$BUILD_NUMBER}' http://172.16.5.13:36000/api/v1/ci/job/sync || true"""
+        sh """curl --connect-timeout 2 --max-time 4 -d '{"job":"$JOB_NAME","id":$BUILD_NUMBER}' http://172.16.5.25:36000/api/v1/ci/job/sync || true"""
     }
 }
